+++ Template/source/karma.conf.js
@@ -3,11 +3,14 @@
 var path = require('path');
 var gulp = require('gulp');
 var conf = require('./gulp/conf');
-
 var paths = gulp.paths;
 
 var _ = require('lodash');
 var wiredep = require('wiredep');
+
+var pathSrcHtml = [
+  path.join(paths.src, '/**/*.html')
+];
 
 function listFiles() {
   var wiredepOptions = _.extend({}, conf.wiredep, {
@@ -15,14 +18,27 @@ function listFiles() {
     devDependencies: true
   });
 
-  return wiredep(wiredepOptions).js
+  var patterns = wiredep(wiredepOptions).js
     .concat([
       path.join(paths.src, '/app/**/*.module.js'),
       path.join(paths.src, '/app/**/*.js'),
       path.join(paths.src, '/**/*.spec.js'),
       path.join(paths.src, '/**/*.mock.js'),
-      path.join(paths.src, '/**/*.html')
-    ]);
+    ])
+    .concat(pathSrcHtml);
+
+  var files = patterns.map(function(pattern) {
+    return {
+      pattern: pattern
+    };
+  });
+  files.push({
+    pattern: path.join(paths.src, '/assets/**/*'),
+    included: false,
+    served: true,
+    watched: false
+  });
+  return files;
 }
 
 module.exports = function(config) {
@@ -34,15 +50,17 @@ module.exports = function(config) {
 
     autoWatch: false,
 
-    frameworks: ['jasmine', 'angular-filesort'],
+    ngHtml2JsPreprocessor: {
+      stripPrefix: paths.src + '/',
+      moduleName: 'app'
+    },
+
+    logLevel: 'WARN',
+
+    frameworks: ['phantomjs-shim', 'jasmine', 'angular-filesort'],
 
     angularFilesort: {
       whitelist: [path.join(paths.src, '/**/!(*.html|*.spec|*.mock).js')]
-    },
-
-    ngHtml2JsPreprocessor: {
-      stripPrefix: 'src/',
-      moduleName: 'gulpAngular'
     },
 
     browsers : ['PhantomJS'],
@@ -50,14 +68,32 @@ module.exports = function(config) {
     plugins : [
       'karma-phantomjs-launcher',
       'karma-angular-filesort',
+      'karma-phantomjs-shim',
+      'karma-coverage',
       'karma-jasmine',
       'karma-ng-html2js-preprocessor'
     ],
 
-    preprocessors: {
-      'src/**/*.html': ['ng-html2js']
+    coverageReporter: {
+      type : 'html',
+      dir : 'coverage/'
+    },
+
+    reporters: ['progress'],
+
+    proxies: {
+      '/assets/': path.join('/base/', paths.src, '/assets/')
     }
   };
+
+  // This is the default preprocessors configuration for a usage with Karma cli
+  // The coverage preprocessor is added in gulp/unit-test.js only for single tests
+  // It was not possible to do it there because karma doesn't let us now if we are
+  // running a single test or not
+  configuration.preprocessors = {};
+  pathSrcHtml.forEach(function(path) {
+    configuration.preprocessors[path] = ['ng-html2js'];
+  });
 
   // This block is needed to execute Chrome on Travis
   // If you ever plan to use Chrome and Travis, you can keep it