+++ Template/source/src/app/app.module.js
@@ -6,9 +6,8 @@
             'ui.router',
             'triangular',
             'ngAnimate', 'ngCookies', 'ngSanitize', 'ngMessages', 'ngMaterial',
-            'googlechart', 'chart.js', 'linkify', 'ui.calendar', 'angularMoment', 'textAngular', 'uiGmapgoogle-maps', 'hljs', 'md.data.table', angularDragula(angular), 'ngFileUpload',
-            // 'seed-module',
-            // uncomment above to activate the example seed module
+            'googlechart', 'chart.js', 'linkify', 'ui.calendar', 'angularMoment', 'textAngular', 'uiGmapgoogle-maps', 'hljs', 'md.data.table', angularDragula(angular), 'ngFileUpload', 'nvd3',
+
             'app.translate',
             // only need one language?  if you want to turn off translations
             // comment out or remove the 'app.translate', line above
@@ -17,6 +16,8 @@
             // comment out or remove the 'app.permission', line above
             // also remove 'permission' from the first line of dependencies
             // https://github.com/Narzerus/angular-permission see here for why
+            // uncomment above to activate the example seed module
+            // 'seed-module',
             'app.examples'
         ])