+++ Template/source/src/app/permission/user.factory.js
@@ -8,8 +8,8 @@
     /* @ngInject */
     function UserService($q, $http, RoleStore) {
         var currentUser = {
-            displayName: 'Christos',
-            username: 'christos',
+            displayName: 'Larry Hilton',
+            username: 'larryhilton',
             avatar: 'assets/images/avatars/avatar-5.png',
             roles: ['SUPERADMIN']
         };