+++ Template/source/src/app/triangular/components/table/table-start-from.filter.js
@@ -11,8 +11,10 @@
         ////////////////
 
         function filterFilter(input, start) {
-            start = +start;
-            return input.slice(start);
+            if (input && input.length > 0) {
+                start = +start;
+                return input.slice(start);
+            }
         }
     }